midwork

+++ import/cppmakelib/utility/time.py
@@ -1,1 +1,1 @@
-time = int
+time = float
